Add trivial support for PMPs, fixes boot on newer qemu
We were previously noncompliant with the RISC-V specification, and so
was qemu: if there are no PMP registers configured, according to the
spec, any access to memory, including instruction fetch, from S- or U-
mode, should fail (see section 3.6.1 of the RISC-V privileged spec).

However, qemu was not enforcing this properly prior to
https://github.com/qemu/qemu/commit/d102f19a208. Running a qemu newer
than this would get stuck in an infinite loop hitting the exception
vector repeatedly.

This commit adds trivial support for the PMP registers, writing a value
such that all accesses from S- and U- mode have their access managed
solely by the page tables. This restores the previous behaviour.

Further, it makes the machine mode timer vector check if it receives an
unexpected interrupt if so, jump to a specific infinite loop. This
should make any potential future occurrences of interrupt issues more
obvious.